01-18-23 v2.9.1 01-18-23 Commit
Front door broke in the 12-31-22 v2.8.1 version of the project.  After a lot of troubleshooting and not being able to spot the issue I fixed this by reverting back to version 12-30-22 and then redoing all of the pieces from 12-31-22 and more.

I re-fixed the door button and the colliders by relinking missing items and moving the colliders back into place.

I found and deleted duplicate scripts and duplicate colliders for the doors that were part of the door prefabs I brought over from the 12-30-22 project and part of the 12-30-22 project itself.

All door items (manually opening and closing, opening and closing correctly using the door toggle button, deactivating the front door collider that prevents the player from running into that door with physics and that offers the same pushback script/mechanic as the walls in my project) all work perfectly again.  Some of those door colliders were in the middle of the room and were preventing the raycasts from reaching the main menu.  Moving these back into place also fixed this issue.

I wrote and added a script for all five cabinet doors and the cabinet drawer to freeze both the x, y and z position and the x, y and z rotation when collisions occur.  My first script used the OnTriggerEnter and OnTriggerExit methods with a tag check but I later changed this over to use the OnCollisionEnter and OnCollisionExit methods so that I didn't have to make my colliders "Is Trigger," which defeats the purpose of having a collider with physics hands and that also couldn't be applied to all parts of the drawer, else items would fall right through it.  I also added a "Wall" tag to all my cabinet doors and drawers so that these also offered the same pushback as the walls and front door when leaning into just a cabinet door or drawer.  This prevents any sort of breaking of the hinge joint or rubber banding effect with collision.

I found my two Unity videos and my Super Mario World video were present but corrupt in the rebuilding of my project to the 01-18-23 build.  So I grabbed those videos from a past build on my external USB HD and I copied them back over to the current project.  I tested all three videos and this worked.

I fixed the tablet black attach point so that this had both the correct orientation (not upside down) but also not close to the body, at the same height as the hand, but also tilted back slightly for easier viewing of the videos on the tablet.

I fixed the "Matt's Photos" red pushpin that was stretched to unrealistic proportions to make this look more like the object it is, a pushpin.

I double checked the Bon Apatite sign to make sure this was readable (I had apparently fixed this prior to my 12-30-22 commit by making this text larger and making sure it wasn't stretched via the scale and adjusting the font color and font properties), but I couldn't find mention of this in my previous log, so I might have forgotten to note that previous change.

I added random sounds to my Dalek figure, just because I could. :-D  When you pick up the Dalek and activate (press the trigger on the VR controller) a random Dalek clip plays (out of ten total).  I was going to write my own script to play random sounds but Unity had included their own in the CreateWithVR course so I just ended up using it instead.

I fixed the attach transform for the dart gun (as it was being picked up and pushed/clipping through where the body would be and tilted upward.  I changed this so it's straight and away from the body.  
I added an attach transform to the Dalek so that when you pick it up, it's facing you.  Pressing the trigger button plays a random clip of about 10.  
 
I fixed the magnifying glass so that it now magnifies either direction.  There's an attach transform on it so you have to sort of bend your wrist to see the other side but before it was like a mirror (showing what was behind you) on the backside instead of being identical to the front side.  I also fixed the magnification so that it magnified more without clipping through the wall and showing outside the room.

I wrote a script using OnCollisionEnter and OnCollisonExit for the watering can so that whenever the watering can collides with one of about four or five tags (which includes Floor, Furniture, and Untagged, and several others) the watering can's X and Z rotation are set back to zero while the Y rotation stays the same.  This is so when the watering can is thrown or set down it doesn't continue to play the audio glugging sound as well as play the water pour particle animation.

I added attach transforms to the Bon Apatite sign and to the Polaroid Camera for proper rotation and position when holding these items.

Every interactable was play tested tonight, as well as the scene changes, and everything is working well in this version.

To Fix Next:

Writing a script for the Water Bottle Flip Challenge to detect when there is a floor collision with either the top, sides or bottom and to play a corresponding sound when that happens.

Determine why my physics hands can clip through the dinning room table and chairs when at a fast enough speed. Try and make this work more like the couch or outside colliders. Test with other furniture to see if anything else in the room allows this to happen. Compare and contrast to troubleshoot.

Thin items like photos from the polaroid camera and my punch list sheets are still sometimes getting stuck under the floor and I need to troubleshoot that more to try and fix.

Updating the Interactive board and/or punch lists.  I'm running out of room on my punch lists and I have no more room to add additional punch lists, so I'll have to decide how I wish to use these moving forward.  I still love the idea of having bugs and/or future features viewable in VR and I have a lot of future ideas (both from ideas I've come up with and that my students have come up with that could be added).  I'm about 1/3 of the way through laying this out but moved onto other parts of the project - just because of the amount of work involved and in not knowing if everything will fit or what the end result of this may look like.

If I get that working I feel like I have a lot of the bug pieces for all the current items added to my room fixed. I'll have to go back to my bug list to check to be sure, but off the top of my head there aren't any others that are coming to mind that I need to fix.  At this point I may choose a new feature to add to this room, just to continue my (and my student's) learning.
@mattdway mattdway committed on Jan 18