02-12-23 v3.0.0 Commit
02-12-23 v3.0.0 Commit
Updated this project to Unity 2021.3.11f1 and I also updated all of the Project Manager plugins to their latest versions including XRIT to version 2.3-pre1.  I fixed a slew of bugs that occured as part of that upgrade and I've done a lot of beta testing.

I also added a new feature where the players can now set the area rug on fire using the lighter as a little easter egg.