05/29/23 Commit
Tweaked spacing of items on counter to make it easier to hit the next button on the welcome screen.

Moved the next button on the welcome screen to the left.

Updated Interaction Board list to include new items.

Changed south wall position to prevent left window from shattering on game start.

Tested to make sure that I couldn't escape the room via any gap I created with the above change.

Play tested the entire room to make sure everything was working correctly.

Fixed candles that had an issue.

Removed duplicate rigidbody in magnifying glass that was causing it to jump out of the player's hand when picked up and to drop out of the room.

Fixed potion bottle drink sound and glass shatter sound.

Changed breakable window settings to make windows less likely to be broken.

Hid brick in cabinet with the blue door.

Started working on gym (added ladders back, added another climbing wall, started to work on scripts.